% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{align_modout}
\alias{align_modout}
\title{Align model output to observed data for comparison}
\usage{
align_modout(modout, obs)
}
\arguments{
\item{modout}{Raw model output}

\item{obs}{Observed data, as returned by \code{\link{get_obsdata}}.  The time
column (= day - day0) must already have been added.}
}
\description{
Align model output to observed data for comparison
}
\keyword{internal}
