% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-filter.R
\name{bayesian_filter}
\alias{bayesian_filter}
\title{Project infection in a single county using a Bayesian filter}
\usage{
bayesian_filter(
  initstates,
  locality,
  tinit,
  tfin,
  obsdata = NULL,
  nsamp = 1000,
  prior = bayes_filter_default_prior
)
}
\arguments{
\item{initstates}{Vector of initial state and parameter samples (see details).}

\item{tinit}{Time corresponding to the initial states.}

\item{tfin}{End time for the forecast.}

\item{obsdata}{Observed number of tests and number of positive tests by
county. (See details)}

\item{nsamp}{Number of Monte Carlo Samples to run for each state.  The final set
of parameters from the Monte Carlo will be used to generate the finals state.}

\item{prior}{Prior PDF function.  It should take a vector of five parameters:  start and end
beta values, start and end importation values, and size of time step.}
}
\value{
A table of parameter values and final states corresponding to each initial
state.  This output is suitable to feed back into the function as \code{initstates}
for a subsequent week.
}
\description{
This function runs a single county one week forward in time from a distribution
of model states.  For each starting state it produces one sample from the posterior PDF
of beta values and importation values for the next week.  The result is a distribution
of one-week beta values and importation values, along with a new distribution of
states at the end of the week.
}
\details{
The initial states should be a matrix with samples in rows and variables in
columns.  The variables should be the state variables of the infection equations
(S, E, I, Is, and R), plus the beta and import_cases parameters.

The observed data should be the first element of the list returned by
\code{\link{get_obsdata}}, filtered to the locality being analyzed.  If
omitted, we will fetch the data and filter it; however, if this function will
be called consecutively for several weeks, running, it will save some time to
do the filtering once and pass the table in.
}
