% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{coefs}
\alias{coefs}
\alias{calcalpha}
\alias{calcbeta}
\alias{calcgamma}
\alias{calcr0}
\title{Caclulate model coefficients from parameters}
\usage{
calcalpha(parms)

calcbeta(parms)

calcgamma(parms)

calcr0(parms)
}
\arguments{
\item{parms}{Named vector of model parameters}
}
\description{
We parameterize the model SEIR in terms of initial doubling time, incubation period, 
and recovery time.  The model equations are given in terms of rates,
\eqn{\alpha}, \eqn{\beta}, and \eqn{\gamma}.  Also of interest is the basic
reproduction parameter \eqn{R_0}.
}
\details{
Recovery rate and progression rate are just the reciprocals of their respective
periods:
\deqn{\alpha = 1/A_0}
\deqn{\gamma = 1/D_0}

Beta is calculated from \eqn{\alpha}, \eqn{\gamma}, and \eqn{t_d} as
\deqn{\beta = \gamma + \ln 2 \frac{(\alpha+\gamma)}{\alpha t_d}}

\eqn{R_0} is calculated from \eqn{\beta} and \eqn{\gamma}.  Note that in the 
absence of mortality during the exposed period, \eqn{R_0} is independent of 
\eqn{\alpha}.
\deqn{R_0 = \frac{\beta}{\gamma}}
}
\section{Functions}{
\itemize{
\item \code{calcalpha}: Calculate progression rate \eqn{\alpha} from model parameters

\item \code{calcbeta}: Calculate transmissibility \eqn{\beta} from model parameters.

\item \code{calcgamma}: Calculate recovery rate \eqn{\gamma} from model parameters.

\item \code{calcr0}: Calculate basic reproduction number \eqn{R_0} from model parameters.
}}

