% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{complete_params}
\alias{complete_params}
\title{Add default parameters for parameters not overridden in the input}
\usage{
complete_params(params)
}
\arguments{
\item{params}{Named list of parameter values}
}
\value{
Named list that gives all parameters an explicit value.
}
\description{
Any parameters not mentioned in the parameter list will have an entry added using the defaults
above.
}
\keyword{internal}
