% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llanalysis.R
\name{cutdate}
\alias{cutdate}
\title{Convert vector of dates into a factor of date ranges}
\usage{
cutdate(dates, breaks)
}
\arguments{
\item{dates}{Vector of dates}

\item{breaks}{Number of date ranges to split into}
}
\description{
Convert the date vector into a numeric vector, then apply \code{\link[base]{cut}},
and rename the levels of the factor to reflect the date ranges included.
}
