% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval-model.R
\name{ensemble_deviance}
\alias{ensemble_deviance}
\title{Compute the deviance for models in an ensemble.}
\usage{
ensemble_deviance(filterfit)
}
\arguments{
\item{filterfit}{Filter-fit object for the ensemble.  The history must be
included in order to calculate the deviance (see
\code{\link{fit_filter}}).}
}
\value{
The input object, with a new entry \code{deviance} added.
}
\description{
The deviance for a model is 2(log L - log L0), where L is the likelihood
value for the model and L0 is the likelihood for the saturated model (a
hypothetical model that gets every observation exactly correct).
}
\details{
Theoretically, we should compute the expected observations by averaging the
prevalence over the week corresponding to the observation, but we get a
pretty good approximation to that by just taking the point estimates at the
end of the week.
}
