% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble-eval.R
\name{ensemble_eval}
\alias{ensemble_eval}
\title{Evaluate ensemble members according to their agreement with observations.}
\usage{
ensemble_eval(modlist)
}
\arguments{
\item{modlist}{List of \code{filter-fit} objects. Each fit is for a single 
locality.}
}
\value{
Table of ensemble ID numbers and log-likelihood values, along with some
diagnostic statistics.
}
\description{
For each ensemble member, compare its predicted number of cases with the
observed number of cases and compute a log-likelihood function.
}
\details{
Right now we're just looking at number of cases, given the effective number
of tests.  Eventually we hope to include the number of UVA hospitalizations
as a constraint.
}
