% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-filter.R
\name{extend_filter_model}
\alias{extend_filter_model}
\title{Extend a filter model with new data.}
\usage{
extend_filter_model(filterfit, newobs, tfin = NULL)
}
\arguments{
\item{filterfit}{Structure of class filter-fit returned by
\code{fit-filter}.}

\item{newobs}{Observed dataset to use to extend the fit.  This dataset should
have some observations at later times than the final time in the input
model. Data from times earlier than the final time in the input model will
be ignored.}

\item{tfin}{Last time to use to fit the new model.  Default is to use the
time of the last observation.}
}
\value{
A \code{filter-fit} structure for which the fit has been extended
  using the new data
}
\description{
This function uses the information in a filter model structure to set up
the input for \code{\link{fit_filter}}.  The observed data is supplemented
with a new dataset, and the fit is continued from the final state stored in
the model.
}
