% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{fill_defaults}
\alias{fill_defaults}
\title{Fill in default parameter values}
\usage{
fill_defaults(parms, defaults)
}
\arguments{
\item{parms}{Named vector of parameters}

\item{defaults}{Named vector of default parameters}
}
\description{
Works on vectors and lists.  Returns a structure of the same type as its first input.
}
