% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter-fit-locality.R
\name{filter_fit_locality}
\alias{filter_fit_locality}
\title{Run the filter model for a single locality}
\usage{
filter_fit_locality(
  locality,
  inputdir,
  Nens = 100,
  inputpfx = "filter-ensemble-candidates-",
  inputsfx = ".rds",
  datemax = NULL,
  history = TRUE
)
}
\arguments{
\item{locality}{Either the name of the locality, or the number in the master 
list.  There are 133 localities.}

\item{inputdir}{Directory containing the input files generated by \code{metrosamp}.
The file name will be constructed by appending the locality name to the input file prefix.}

\item{Nens}{Number of ensemble members.}

\item{inputpfx}{Stem of input file name.  Default is \code{'filter-ensemble-candidates-'}.
The name of the locality will be appended to this, followed by the input file suffix.}

\item{inputsfx}{Suffix appended to the input filename.  Default is \code{'.rds'}.}

\item{datemax}{Maximum date to use in the observed data.  Default is to exclude
the last week if it is incomplete (i.e., daily observations aren't current through
the end of the week).}

\item{history}{Flag indicating whether to keep full results from all ensemble members.
If \code{FALSE}, keep only median and confidence intervals at each time step,
plus final state}
}
\description{
This function requires a \code{metrosamp} save file with parameter samples
for the locality.  An ensemble of models is initialized from these samples,
and the filter is run for each ensemble member.
}
