% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagplots.R
\name{filter_model_comparison_plot}
\alias{filter_model_comparison_plot}
\title{Compare filter model fits in various localities}
\usage{
filter_model_comparison_plot(filterfits, what = "fi")
}
\arguments{
\item{filterfits}{List of filter-fit objects}

\item{what}{Name of the variable to plot (fi, beta, ncase, import_cases)}
}
\description{
Compare filter model fits in various localities
}
