% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagplots.R
\name{filter_model_diagnositc_plots}
\alias{filter_model_diagnositc_plots}
\title{Generate diagnostic plots for filter models}
\usage{
filter_model_diagnositc_plots(modelfit)
}
\arguments{
\item{modelfit}{Modelfit element of the filter-fit structure}
}
\value{
A list of ggplot objects
}
\description{
Generate diagnostic plots for filter models
}
