% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-filter.R
\name{fit_filter}
\alias{fit_filter}
\title{Fit a Bayesian filter over time}
\usage{
fit_filter(
  initstates,
  obsdata,
  tinit,
  tfinal,
  hospdata = NULL,
  history = TRUE,
  ids = NULL
)
}
\arguments{
\item{initstates}{Matrix containing ensemble of initial states and parameter
values (see description in \code{\link{bayesian_filter}}.}

\item{obsdata}{Observed data to use in the fit}

\item{tinit}{Initial time corresponding to the states provided in
\code{initstates}}

\item{hospdata}{Hospitalization data to use in the fit.}

\item{history}{Detailed history for the individual ensemble members (see
details).}

\item{ids}{List of serial numbers for the ensemble members.  This is only
used if we are keeping history.  See details for why you might want to
specify this explicitly.}

\item{tfin}{Final time (i.e., the time of the last observation to use in
fitting the filter)}
}
\description{
This function calls \code{\link{bayesian_filter}} repeatedly to fit an
evolving set of parameters to a dataset.  The result is organized into a
table of model state and parameters over time.
}
\details{
The \code{history} argument allows us to optionally keep detailed histories
for the individual ensemble members, including beta, imports, I, Is, and
Itot.  Otherwise, the only history you get is is summary statistics (median
and 95% confidence bounds) for those quantities.  To get the full history,
either pass in the history output from a previous run (the new history will
be appended), or if earlier history is not available, set to any non-null
value (the new history will be returned).

If history is being kept, then we need to assign serial numbers to the
ensemble members.  By default, these are assigned sequentially, but there are
reasons why they might need to be specified explicitly.  One is if the
ensemble matrix is reordered for some reason.  The other is if some ensemble
members are dropped or replaced.
}
