% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{fsympto}
\alias{fsympto}
\title{Find the symptomatic infection fraction}
\usage{
fsympto(modrun, trange = c(80, 93))
}
\arguments{
\item{modrun}{Model output from \code{\link{run_scenario}}}

\item{trange}{Range of times to average symptomatic fraction over.  The endpoints
are included.}
}
\description{
The fraction of infections that are symptomatic varies over the course of the
scenario, so you must pick a range of values to average over.  If no range is
specified, the default is days 80-93 (inclusive) from the start of the simulation.
}
