% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gen_likelihood}
\alias{gen_likelihood}
\title{Prepare a likelihood function for use in Bayesian calibration}
\usage{
gen_likelihood(
  hparms,
  fixed_parms = NULL,
  maxdate = NULL,
  verbose = FALSE,
  waicmode = FALSE
)
}
\arguments{
\item{hparms}{Hyperparameters for the calculation, as described in \code{\link{gen_post}}}

\item{fixed_parms}{A named vector of parameters to use as fixed values for parameters
not passed to the likelihood function.  Note these "fixed" parameters can still
be overridden by passing an explicit value.}

\item{maxdate}{Last date to use in the calibration.  Default is 2020-06-30}

\item{verbose}{If \code{TRUE}, print additional diagnostic information.}

\item{waicmode}{If \code{TRUE}, generate a function that returns the likelihood
contributions from each data point.}
}
\value{
A function that takes a named list of model parameters and returns
a log-likelihood value.  See details for the parameters recognized and the
output of the function returned.
}
\description{
Build a function that given a set of model parameters runs the model and
computes the log-likelihood for the model.
}
\details{
The observational data is the daily county-by-county reports of new confirmed
COVID-19 cases.  We are using the New York Times dataset
(\code{https://github.com/nytimes/covid-19-data}), which is based on reports
from state departments of health, including the VDH.  For statistics on COVID-19
testing, we use the \code{\link[vacovdata]{vacovdata}} dataset, which was compiled
by the COVID Tracking project (\code{https://covidtracking.com/}).

The model parameters currently recognized are:
\describe{
\item{eta}{(real) Log of the baseline transmissibility}
\item{xi}{(real) Coefficient of population density in log-transmissibility}
\item{D0}{(real > 0) The initial average infection duration}
\item{A0}{(real > 0) The average incubation time}
\item{Ts}{(real > 0) The average time from infection to symptom onset.  Note that
this parameter is probably not identifiable with our current data.}
\item{b}{(real > 0) The testing bias factor.  That is, the positive test rate
divided by the true infection rate.  b can be different from 1 because of false positives
or because testing is targeted to people suspected of having the disease.  (Generally,
we expect b>1, but we don't require this.)}
\item{I0}{(real > 0) The initial number of infected people, once the infection starts.
This is taken to be the same in all counties.}
}
For the time being, we start tracking the infection in Fairfax county on day 30, and
all of the counties are delayed relative to Fairfax by a number of days equal
to the difference between their first observed case and the first case observed
in Fairfax.

To compute the likelihood, our assumption is that the output of the model
represents an average infection rate.  The average observation rate is then
\deqn{\bar{N}_o = \bar{N}_I f_t b,} where \eqn{f_t} is the fraction of the total
population tested.  We then assume that the actual observations are distributed
as \deqn{N_o \sim Pois(\bar{N}_o).}
}
\section{Likelihood function}{


By default, the function produced will return a single value, which is the sum of
\itemize{
\item{The log-likelihood function, summed over all data points}
\item{A correction for the hospitalization fraction}
\item{A correction for the symptomatic fraction}
}
Technically, the latter two are priors, since they don't depend on the observations,
but the model must be run in order to compute them, so they are computed here.

If the \code{waicmode} flag is set, then the function will return detailed information
on the contribution of each data point to the log-likelihood.  The result will be a
data frame containing
\itemize{
\item{date}
\item{locality}
\item{expected counts}
\item{observed counts}
\item{Hypergeometric parameters x, m, n, and k}
\item{log likelihood}
}
Additionally, the antibody prevalence and symptomatic fraction corrections
will be attached to the data frame as attributes (\code{hfadjust} and
\code{fsadjust}).

The default mode output can be used in optimizations or Markov chain Monte Carlo.
The waicmode output is useful for computing the WAIC, or for diagnosing which
counties and/or days are favoring one model over another.
}

