% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{gen_parm_vec}
\alias{gen_parm_vec}
\title{Generate a vector of parameters for the posterior PDF}
\usage{
gen_parm_vec(localities = NULL, beta = NULL, common_parms = NULL)
}
\arguments{
\item{localities}{Vector of localities to use in the calculation.  If
omitted, use all localities for which there is case data.}

\item{beta}{Value to use for transmissibility.  All localities will get the
same value.}

\item{common_parms}{Values to use for the statewide parameters.  If omitted,
default values will be used}
}
\value{
Named vector of parameter values.
}
\description{
Create a vector of parameters for use with the posterior function.  All
counties are initialized with the same beta value, and values can be
specified for the statewide parameters.  Any parameters not specified will be
given default values.
}
