% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gen_post}
\alias{gen_post}
\title{Prepare a posterior log-pdf function for use in Bayesian calibration}
\usage{
gen_post(prior_weight = NULL, fixed_parms = NULL)
}
\arguments{
\item{prior_weight}{Factor to multiply the prior by.  Default is to make it equal
to the number of counties for which we have confirmed COVID-19 cases.}

\item{fixed_parms}{A named vector of parameters to use as fixed values for parameters
not passed to the likelihood function.  Note these "fixed" parameters can still
be overridden by passing an explicit value.}
}
\description{
Prepare a posterior log-pdf function for use in Bayesian calibration
}
