% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{gen_prior}
\alias{gen_prior}
\title{Generate a log prior function}
\usage{
gen_prior(hparms, verbose = FALSE)
}
\arguments{
\item{hparms}{Named vector of hyperparameters, described in \code{\link{gen_post}}}

\item{verbose}{If \code{TRUE}, print out additional diagnostic information.}
}
\value{
A function that takes a named vector of parameters and returns the log-prior
probability density
}
\description{
This function returns a a function that given an input set of parameters evaluates
the log-prior probability density for the parameters.  Eventually this function
will be able to take hyperparameter arguments that influence the details of the
priors, but for now they're fixed.
}
\details{
The parameters currently recognized are described in \code{\link{gen_likelihood}}.
If there are any parameters not in that list present, an error will be raised.
}
