% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gen_simobs_likelihood}
\alias{gen_simobs_likelihood}
\title{Generate a likelihood function for single-county simulated data}
\usage{
gen_simobs_likelihood(simobs, timerange, verbose = FALSE)
}
\arguments{
\item{simobs}{Simulated observations, as returned by \code{\link{simobs}}.}

\item{timerange}{Length-2 vector giving the earliest and latest dates to use
in the calculation.  If not specified, use all times in the dataset.}

\item{verbose}{If \code{TRUE}, output additional diagnostic information to
console.}
}
\description{
This likelihood function is greatly simplified from the one produced by
\code{\link{gen_likelihood}}.  It contains none of the adjustments for
hospitalizaton fraction (since simulated data has no hospitalization to
compare to), and it doesn't (yet) distinguish between symptomatic and
asymptomatic cases.
}
\details{
The parameters to this model are beta, A0, D0, Ts, I0, b0, b1, mask_effect,
and import_cases.  Mask effect can be omitted; if so, it is fixed at
zero.

The scenario runs are assumed to start at the earliest time in the observed
dataset, even if the time range for comparing to data starts later.  The
population is initialized to \code{I0} exposures at that time with the rest
of the population initialized as susceptible.

Another change in this function is that we use the binomial distribution
instead of the hypergeometric.  This saves us some mucking about with
ensuring that the susceptible population is greater than the number of
positive observations.
}
