% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gen_simobs_posterior}
\alias{gen_simobs_posterior}
\title{Generate log-posterior function for filter model}
\usage{
gen_simobs_posterior(simobs, timerange, verbose = FALSE)
}
\arguments{
\item{simobs}{Simulated observations (see
\code{\link{gen_simobs_likelihood}})}

\item{timerange}{Time range to use in comparing to observations}

\item{verbose}{Flag to turn on verbose mode.}
}
\description{
Generate log-posterior function for filter model
}
