% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{get_obsdata}
\alias{get_obsdata}
\title{Prepare the observed data for use in the likelihood function}
\usage{
get_obsdata(maxdate = as.Date("2020-12-31"))
}
\arguments{
\item{maxdate}{Maximum date to include in the comparison data.}
}
\value{
List of observations used in the likelihood function.  This includes
\describe{
  \item{obsdata}{Confirmed new cases by date and county.  Also includes 
  an estimate of the number of people tested in each county each day.}
  \item{fips_codes}{Lookup table of FIPS codes for the Locality names output by
  the model.}
  \item{default_parm_vals}{Default values for the parameters used to compare
  model output to observed data (the parameters used inside the model have their
  own defaults)}
}
}
\description{
Prepare the observed data for use in the likelihood function
}
\keyword{internal}
