% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{growth_categories}
\alias{growth_categories}
\alias{hi_counties}
\alias{lo_counties}
\alias{uhi_counties}
\alias{ulo_counties}
\title{County growth categories}
\format{
Data frame with 2 columns
\describe{
\item{fips}{FIPS code}
\item{locality}{Name, using the vdh conventions}
}

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 48 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 34 rows and 2 columns.

An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 16 rows and 2 columns.
}
\usage{
hi_counties

lo_counties

uhi_counties

ulo_counties
}
\description{
COVID-19 growth across the Commonwealth of Virginia has been very heterogeneous.
Over the month of April doubling times have ranged from under 4 days to nearly
40 days.  The growth statistics over April have been used to sort the counties
into four categories.  In the model, each of these categories will get its own
growth rate.
}
\section{Functions}{
\itemize{
\item \code{hi_counties}: High growth rate counties

\item \code{lo_counties}: Low growth rate counties

\item \code{uhi_counties}: Ultra-high growth rate counties

\item \code{ulo_counties}: Ultra-low growth rate counties
}}

\keyword{datasets}
