% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hospitalizations}
\alias{hospitalizations}
\title{Weekly UVAHS Covid admissions by date and locality of origin}
\format{
Data frame with 4 columns
\describe{
\item{date}{Date of last day of the week being reported}
\item{fips}{FIPS code for the locality of origin}
\item{locality}{Name of locality of origin}
\item{hospitalizations}{Number of hospitalizations from that locality for the week}
}
}
\source{
UVAHS records
}
\usage{
hospitalizations
}
\description{
Weekly COVID cases at the HS tabulated by date and locality.  Dates of cases been
adjusted by the average time from symptom onset to hospitalization.
}
\details{
The \code{date} variable is the date of the last day of the week.  Weeks are
defined to end on Sundays.
}
\keyword{datasets}
