% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-filter.R
\name{initialize_parmset}
\alias{initialize_parmset}
\title{Run compartment model from an initial state}
\usage{
initialize_parmset(parms, obsdata, t1, dt = 1)
}
\arguments{
\item{parms}{Named vector of model parameters: beta, A0, D0, Ts, I0, and
mask_effect.}

\item{obsdata}{Data frame of observed data.  The only columns used in this
case are time and total population.}

\item{t1}{End time for the scenario.}

\item{dt}{Time step for scenario output.  Default is dt=1.  Note that if
\code{dt > 1}, the end time \code{t1} might not be included in the output.}
}
\description{
This is a wrapper for \code{\link{run_parmset}} for cases where we have a set
of parameters including an initial infection level, and we want to run it
from the start of a scenario to a specified time.
}
