% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{is.Scenario}
\alias{is.Scenario}
\title{Test whether an object is a scenario}
\usage{
is.Scenario(obj)
}
\arguments{
\item{obj}{Object to be tested}
}
\description{
This function is a convenience wrapper around \code{inherits}
}
