% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llanalysis.R
\name{llanalysis}
\alias{llanalysis}
\alias{ll_bycounty}
\alias{ll_bydate}
\alias{ll_bydaterange}
\alias{cmpll_bycounty}
\alias{cmpll_bydaterange}
\title{Functions for analyzing data contributions to the likelihood}
\usage{
ll_bycounty(lldetail)

ll_bydate(lldetail)

ll_bydaterange(lldetail, n = 5)

cmpll_bycounty(lldetail, lldetail2)

cmpll_bydaterange(lldetail, lldetail2, n = 5)
}
\arguments{
\item{lldetail}{Data frame of detailed log-likelihood contributions.}

\item{n}{Number of ranges to divide dates into}

\item{lldetail2}{Data frame of detailed log-likelihood contributions for a 
comparison scenario.}
}
\description{
These functions all take the table produced by the \code{waicmode} of 
\code{\link{gen_likelihood}} and summarize it by different groupings.
This allows us to see whether certain counties or certain time periods
are having an outsized contribution to the overall likelihood.
}
\details{
To use these functions, run the parameters through a likelihood function
generated by \code{\link{gen_likelihood}} with \code{waicmode = TRUE}.  Then
run the parameters you wish to investigate through that likelihood function
and pass the result as \code{lldetail}.
}
\section{Functions}{
\itemize{
\item \code{ll_bycounty}: Report log-likelihood contributions by county

\item \code{ll_bydate}: Report log-likelihood contributions by date

\item \code{ll_bydaterange}: Report log-likelihood contributions by date ranges

\item \code{cmpll_bycounty}: Compare log-likelihood contributions by county for two scenarios.

\item \code{cmpll_bydaterange}: Compare log-likelihood contributions by date range for two scenarios
}}

