% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{localbeta}
\alias{localbeta}
\title{Calculate the local value of beta from population density}
\usage{
localbeta(parms, localities)
}
\description{
Our formula is \eqn{\beta = \exp(\eta + \xi d)}, where \eqn{d} is the 
standardized population density.
}
\details{
This function is vectorized over localities.
}
