% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{locality_startdate}
\alias{locality_startdate}
\title{Compute start of the infection for a locality.}
\usage{
locality_startdate(localities)
}
\arguments{
\item{localities}{Names of localities to get dates for.}
}
\description{
The infection is deemed to start in each locality at a fixed date (set by the
\code{infection_t0} constant) plus an offest equal to the difference in days
between the first observed case in the locality and the first observed case
statewide.
}
