% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{localize_scenario}
\alias{localize_scenario}
\title{Specialize a scenario to a locality}
\usage{
localize_scenario(scen, locality)
}
\arguments{
\item{scen}{A \code{\link{Scenario}} object.}

\item{locality}{Name of the locality to specialize to.}
}
\value{
A \code{LocalScen} structure.  This object is no longer a scenario
  object because it can no longer be used with an arbitrary locality.
}
\description{
Find the entries in a scenario structure that are germane to a particular
locality.  This saves having to filter by locality every time you go to look
up an adjustment value.
}
