% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{losig}
\alias{losig}
\title{Compute a sigmoid function of the log-odds of a probability}
\usage{
losig(p, p0 = 0.5, sig = 5, log = TRUE)
}
\arguments{
\item{p}{Vector of probabilities \eqn{0 \le p \le 1}.}

\item{p0}{Threshold probability value.}

\item{sig}{Steepness of the cliff at the threshold.  Higher values produce a
steeper dropoff.  Values greater than zero favor probabilities above the
threshold; values less than zero favor probabilities below the threshold.}

\item{log}{If \code{TRUE}, return the log penalty.  Unlike probability
density functions, this flag is set \code{TRUE} by default.}
}
\description{
This function serves as a likelihood adjustment for probabilities where we
believe that the value should be greater than (or less than) some cutoff, but
we don't have a preference for where within the range above (or below) the
threshold the value should be.
}
\details{

}
