% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marketFractionFinal}
\alias{marketFractionFinal}
\title{Market fractions for Virginia localities}
\format{
Data frame with 3 columns
\describe{
\item{Locality}{Name of the locality}
\item{TypeAMCDRG}{Synthetic key combining the type (always 'fraction'), DRG (either 'All' or 'Resp'),
and medical center (either 'UVA' or 'VCU')}
\item{marketShare}{The medical center's market share in that locality, for that DRG, expressed as a 
fraction}
}
}
\source{
(not given)
}
\usage{
marketFractionFinal
}
\description{
Dataset giving the market fractions by locality (county or county-equivalent) for UVA and VCU, 
for all DRGs and for the Respiratory DRG
}
\keyword{datasets}
