% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{mask_indicator}
\alias{mask_indicator}
\alias{mask_indicator.default}
\alias{mask_indicator.Date}
\title{Indicator function for mask use requirement}
\usage{
mask_indicator(t)

\method{mask_indicator}{default}(t)

\method{mask_indicator}{Date}(t)
}
\arguments{
\item{t}{Vector of dates or times (days since Jan 01)}
}
\description{
For an input vector of dates (or times given as days since 2020-01-01), return 
1 for dates on which mask use was required, zero for dates on which it was not.
}
\details{
Executive order 63 established a requirement to wear masks in most public 
places in the Commonwealth of Virginia, starting Friday, 29 May 2020.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default version

\item \code{Date}: Date class version
}}

