% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{mobility_adjust}
\alias{mobility_adjust}
\title{Compute transmissibility adjustment for mobility}
\usage{
mobility_adjust(t, zeta, mobility_table)
}
\arguments{
\item{t}{Days since 2020-01-1}

\item{zeta}{Model coefficient for influence of mobility on transmissibility}

\item{mobility_table}{Mobility table for the locality being processed (see details)}
}
\description{
Compute multiplicative adjustment to transmissibility, based on percentage
a population mobility measurement
The adjustment is \code{exp(zeta * mobility)}
}
\details{
The mobility table is processed in \code{\link{local_mobility}}.
}
