% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{namebackup}
\alias{namebackup}
\title{Find a name for a backup file}
\usage{
namebackup(name)
}
\arguments{
\item{name}{Name(s) of the files to back up}
}
\description{
The backup name is formed by first cutting off the extension (everything after
the final '.') and adding the tag ```_NNN_```, where N counts up from the
number of the last backup.  Then the extension is added back.
}
