% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{nhosp_likelihood}
\alias{nhosp_likelihood}
\title{Calculate the likelihood adjustment for number of hospital observations}
\usage{
nhosp_likelihood(alpha, beta, modout)
}
\arguments{
\item{alpha}{First shape parameter in the prior for hospitalization fraction}

\item{beta}{Second shape parameter in the prior for hospitalization fraction}

\item{modout}{Model output from \code{\link{run_scenario}}}
}
\description{
The adjustment is 
\deqn{L_h = \log(\int dp_h Beta(\alpha_h, \beta_h) \prod Pois(N_h, \lambda(p_h)))}
}
