% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{padjust}
\alias{padjust}
\title{Adjust probabilities by a bias factor}
\usage{
padjust(p, b)
}
\arguments{
\item{p}{Vector of probability values}

\item{b}{Bias factor}
}
\description{
Given a vector of probabilities and a bias factor (or vector of bias factors), 
compute the new probabilities adjusted for the bias factor.
}
\details{
The bias factor acts as a multiplier on the odds ratio, so 
\deqn{\frac{p'}{1-p'} = b \frac{p}{1-p}.}  Thus, for example, if \eqn{p=0.1}, 
the odds ratio is \eqn{r=1/9}.  A bias factor of 10 would produce \eqn{r'=10/9},
for an ajusted probability of \eqn{p' \approx 0.53}.
}
