% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\docType{data}
\name{parameters}
\alias{parameters}
\alias{default_hparms}
\alias{fill_defaults}
\title{Model parameters and hyperparameters}
\format{
\code{default_hparms} : Named list of default values for prior and likelihood
hyperparameters
}
\usage{
default_hparms

fill_defaults(parms, defaults)
}
\arguments{
\item{parms}{Named vector of parameters}

\item{defaults}{Named vector of default parameters}
}
\description{
Add default values for any values not given in the first argument.
Works on vectors and lists.  Returns a structure of the same type as its first input.
}
\keyword{datasets}
