% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagplots.R
\name{plt_modobs}
\alias{plt_modobs}
\title{Plot model predictions against data observations}
\usage{
plt_modobs(parms, ncounty = 4)
}
\arguments{
\item{parms}{Parameters to run model for}

\item{ncounty}{Number of counties to include in the plot}

\item{obs}{Observed data as returned by \code{\link{get_obsdata}}}
}
\description{
Plot model outputs and observed data.
}
