% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-filter.R
\name{project_filter_model}
\alias{project_filter_model}
\title{Continue a run from a fitted filter model.}
\usage{
project_filter_model(filterfit, tfinal, dt = 1, tvintage1 = NA, dtvintage = 7)
}
\arguments{
\item{filterfit}{Structure of class filter-fit returned by
\code{\link{fit_filter}}.}

\item{tfinal}{Final time for the projection.}

\item{dt}{Time step in days.  Note that if \code{dt > 1}, then \code{tfinal}
will not necessarily be in the final output.}

\item{tvintage1}{Time for the first projection vintage.  Having multiple
vintages is only possible if history information was included in
\code{filterfit}.}

\item{dtvintage}{Time step between vintages.}
}
\value{
Data frame with projections in it.
}
\description{
Unlike \code{\link{fit_filter}}, this function just continues running the
model forward in time, with no parameter updates.  Final state and summary
statistics are computed and appended to the ones computed for the original
fit.  History is appended if it is present in the original fit.
}
