% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-mc.R
\name{run_mc}
\alias{run_mc}
\title{Run a Monte Carlo simulation of a scenario}
\usage{
run_mc(
  N,
  distparms,
  tmax,
  baseparms = list(),
  scenarioName = "HospCensus",
  rawsims = FALSE
)
}
\arguments{
\item{N}{number of MC iterations to run}

\item{distparms}{Parameter distributions.  See details for how to specify a
distribution.}

\item{tmax}{End time for simulations}

\item{baseparms}{Base parameters.  Any parameters without distributions will
be set to the values specified here (or to their default values, if unspecified.)}

\item{rawsims}{Flag indicating whether to return the raw simulation output, or
a summarized version.}
}
\value{
A list of data frames, one for each of the forecast variables returned by 
\code{\link{run_scenario}.  The form of the outputs depends on the setting
of the \code{rawsims} input.  If \code{rawsims} is \code{FALSE}, then we get back
a data frame with ensemble mean, median, standard deviation, 5th, and 95th percentiles,
reported by time.
If \code{rawsims} is \code{TRUE}, then each output is a data frame with the output
variable reported by time and MC iteration.}
}
\description{
Run a Monte Carlo simulation for a scenario given fixed distributions for one
or more of the simulation parameters.  At each iteration all of the
parameters for which distributions were supplied will have values selected
from their distributions, while other parameters will be held at the base
values supplied.
}
\details{
The distparams input should be a named list of functions that return samples
from the indicated variables.  For example, if we wanted to have \code{hospFrac}
distributed as Beta(4, 98) and \code{T0} distributed as Gamma(20, 4), we would
write:
\preformatted{
distparms <- list(
   T0 = function(N) {rgamma(N, 20, 4)},
   hospFrac = function(N) {rbeta(N, 4, 98)})
}
Helper functions such as \code{\link{genbeta}} provide an easy way to generate
functions for common distributions.  Using the helper functions you would write
these distributions as:
\preformatted{
distparams <- list(
    T0 = gengamma(20, 4),
    hospFrac = genbeta(4, 98))
}
}
