% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{run_parms}
\alias{run_parms}
\title{Run the scenario for a vector of likelihood parameters}
\usage{
run_parms(parms, scenario_name = "Scen", tmax = 273, aggregate = FALSE)
}
\arguments{
\item{parms}{Vector of parameters used in the likelihood function (q.v.,
\code{\link{gen_likelihood}}).  Unlike with the likelihood function, there
is no provision for default parameters}

\item{scenario_name}{Optional name for the scenario}

\item{tmax}{Maximum time to run to (default is 273, which is 2020-09-30)}

\item{aggregate}{If \code{TRUE}, aggregate across counties at each time.
Otherwise, return the unaggregated data.}
}
\description{
This function extracts the parameters that need to be passed to
\code{\link{run_scenario}}, figures out the time values to use,
and filters out the fractional days from the output.
}
\seealso{
\code{\link{run_scenario}}
}
