% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{run_scenario}
\alias{run_scenario}
\title{Run a single hospital census scenario}
\usage{
run_scenario(
  timevals,
  params = list(),
  counties = NULL,
  scenarioName = "communityInfection"
)
}
\arguments{
\item{timevals}{Times to output values for.  If only a single value is provided, 
it is assumed to be a maximum time (days) to run to, and values will be output
in one day steps from 0 to tmax.}

\item{params}{List of parameter values, see details}

\item{counties}{If provided, run just these counties}

\item{scenarioName}{Name for the scenario; this will be copied into results}
}
\value{
Data frame with results (see details) over time
}
\description{
Run a hospital census scenario for a single set of model parameters.  Results for several
projected variables will be returned in a single data frame.
}
\details{
The results returned will be:
\itemize{
\item{New community infections in the selected counties}
\item{New symptomatic community infections in the selected counties}
\item{Total infected population in the selected counties}
\item{Symptomatic infected population in the selected counties}
\item{Recovered population in the selected counties}
\item{Remaining susceptible population in the selected counties}
\item{Cumulative infected population (includes those who later recovered)}
\item{Cumulative infected share of population (includes recovered)}
\item{UVA market share of symptomatic infections}
}

The parameters for the model are:
\describe{
\item{eta}{(scalar) base value for initial transmissivity}
\item{xi}{(scalar) coefficient for population density in initial transmissivity}
\item{D0}{(scalar) Base recovery time.  Average base recovery time for an infected person.}
\item{A0}{(scalar) Base incubation time.  Average tie for an exposed person to 
become infected.}
\item{(various time lags)}{Time after infection for various events (RTS for details)}
\item{(a whole slew of others)}{see definition of \code{param_defaults} for definitions}

}
}
\seealso{
\code{\link{run_parms}}
}
