% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{run_single_county}
\alias{run_single_county}
\title{Run the hospital census model for a single locality}
\usage{
run_single_county(locality, mktshare, timevals, params, hicounties)
}
\arguments{
\item{locality}{Name of the locality}

\item{mktshare}{Market share for the locality}

\item{timevals}{Vector of output times for the simulation}

\item{params}{Model parameter list}

\item{hicounties}{List of counties that should use the high growth rate.}
}
\description{
Run the hospital census model for a single locality
}
\keyword{internal}
