% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampleCounties}
\alias{sampleCounties}
\title{Market fraction and FIPS code for Virginia localities with high UVA market fractions}
\format{
An object of class \code{data.frame} with 11 rows and 4 columns.
}
\usage{
sampleCounties
}
\description{
Same as \code{\link{marketFractionFinal}}, except includes an additional column for FIPS
code and hs been filtered to just localities with UVA market shares >= 20%
}
\keyword{datasets}
