% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{Scenario}
\alias{Scenario}
\title{Construct a scenario structure}
\usage{
Scenario(scen)
}
\arguments{
\item{scen}{Input data frame describing the scenario.  See details.}
}
\value{
A scenario structure.
}
\description{
A scenario structure is a list of data frames.  Each list element pertains to
one model variable.  The data frame for the variable gives the change in the
parameter by time and locality.
}
\details{
The input data frame should include the following columns:

\describe{
\item{locality}{The locality the change applies to.  A \code{NA} value here
means the change is a default that will apply to any locality without a
specified change.}
\item{time}{The time at which a change in a parameter occurs.  The parameter
will remain at its changed value until another change is imposed.}
\item{parm}{Parameter affected by the change}
\item{isrelative}{Flag indicating whether the change is a multiplier
(\code{TRUE}) or an adder (\code{FALSE}) to the base value.}
\item{value}{The change factor.  This will be treated as an adjustment to the
parameter's value at the end of the historical period.}
}
}
\section{Notes}{

Although the relative flag is given at each time step (due to the limitations
  of the data frame structure), having an adjustment
  that is relative at some times and absolute at others is not supported.
  A mixed vector for the \code{isrelative} flag will result in the adjustment
  being treated as absolute (i.e., additive).

Changes for times in the past (prior to the start of the projection) are
ignored; however, the most recent such change will take
effect in the first future time step.  Thus, one can keep using the same
  scenario structure week after week as the model gets updated.
}

