% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{scenario_change_times}
\alias{scenario_change_times}
\title{Find the times at which a scenario's parameters change}
\usage{
scenario_change_times(localscen)
}
\arguments{
\item{localscen}{A \code{LocalScen} object}
}
\value{
Vector of times at which at least one parameter changes.
}
\description{
Find the change times (i.e., the times at which at least one parameter
changes) across all parameters.
}
