% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scenarios.R
\name{scenario_parm_value}
\alias{scenario_parm_value}
\title{Compute parameter value by time for a local scenario}
\usage{
scenario_parm_value(locscen, timevals, parmbase)
}
\arguments{
\item{locscen}{A \code{LocalScen} object}

\item{timevals}{Times to retrieve parameter values for}

\item{parmbase}{Named vector of base values of the parameters}
}
\value{
Matrix of modified parameter values, each variable is in a column,
  each time in a row.
}
\description{
Compute parameter value by time for a local scenario
}
