% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{select_T0}
\alias{select_T0}
\title{Select the appropriate T0 parameter for a locality}
\usage{
select_T0(locality, parms)
}
\arguments{
\item{locality}{Name of the locality}

\item{parms}{Named list of parameters.  It must have entries for 'T0_uhi',
'T0_hi', 'T0_lo', and 'T0_ulo'}
}
\description{
See discussion in \code{\link{growth_categories}}
}
\keyword{internal}
