% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/obs.R
\name{simobs}
\alias{simobs}
\title{Create simulated observations from a scenario run}
\usage{
simobs(rundata, bparms, ntestfn = linear_ntest)
}
\arguments{
\item{rundata}{Matrix of run data returned from an integration of 
\code{\link{seir_equations}}}

\item{bparms}{Vector containing the b0 and b1 parameters.}

\item{ntestfn}{Function of time that returns the number of daily tests performed.}
}
\description{
Given a matrix of simulation outputs and a function of time for the number
of tests, create simulated observations.
}
