% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sir_equations}
\alias{sir_equations}
\title{Differential equations for the SIR model}
\usage{
sir_equations(time, variables, parameters)
}
\arguments{
\item{time}{Simulation time value}

\item{variables}{Vector of current variable values (see details)}

\item{parameter}{Vector of parameter values (see details)}
}
\value{
A list, as described in \code{\link[deSolve]{ode}}.  In this case we provide only
the first element of the list, which is a vector of derivative values.
}
\description{
Computes the derivatives of the susceptible (S), infected (I), and recovered (R) 
populations.
}
\details{
The variables for this model are S, I, and R, the susceptible, infected, and recovered 
populations.  They should be passed as a named vector, in that order.

The parameters for the model are beta, the infection parameter, and gamma, the recovery
parameter.  These should be passed in as a named vector; the order doesn't matter.
}
