% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayes-filter.R
\name{update_filter_models}
\alias{update_filter_models}
\title{Update all saved filter models}
\usage{
update_filter_models(indir, runloc = NULL, outdir = NULL, save = TRUE)
}
\arguments{
\item{indir}{Directory with input files}

\item{runloc}{Run just the indicated localities.  Can be either a list of
indices or a list of locality names.}

\item{outdir}{Directory to save output in.  Default is
\code{filter-updates.<date>}, where \code{<date>} is the date of the last
observations in the dataset used in the fits.}

\item{save}{Flag indicating whether to save new model fits to disk.}
}
\value{
List of filter models with the fit extended to the new data
}
\description{
Load all the filter models in an input directory and update them with the
latest data.  Write results to a new directory.
}
