% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{uva_covid_count}
\alias{uva_covid_count}
\title{Counts of various categories of COVID cases}
\format{
Data frame with 10 columns
\describe{
\item{date}{Date of observation}
\item{COVID}{Total number of COVID-19 cases}
\item{Admits}{Number of COVID-19 cases admitted that day}
\item{DCs}{Discharges}
\item{ICU}{Number of COVID-19 cases in ICU}
\item{ICU in}{Number of new ICU cases}
\item{ICU out}{Number of cases leaving the ICU}
\item{Vent}{Number of cases on IMV}
\item{vent in}{Number of new IMV cases}
\item{vent out}{Number of IMV cases leaving}
}
}
\usage{
uva_covid_count
}
\description{
Counts of various categories of COVID cases
}
\keyword{datasets}
