% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{va_county_first_case}
\alias{va_county_first_case}
\title{Date of first report for VA counties}
\format{
Data frame with 4 columns
\describe{
\item{FIPS}{FIPS code for the county.}
\item{firstReport}{Date of the first reported COVID-19 case in the county.}
\item{firstDay}{Number of days since 1Jan2020 for the firstReport}
}
}
\source{
New York Times COVID-19 dataset \url{https://github.com/nytimes/covid-19-data}
}
\usage{
va_county_first_case
}
\description{
Date of the first reported COVID-19 case for each Virginia county.  Counties
that have not yet had a reported case have a date of \code{NA}.
}
\keyword{datasets}
