% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{va_county_ntest}
\alias{va_county_ntest}
\title{Estimated number of tests by locality and date}
\format{
Data frame with 6 columns
\describe{
\item{date}{Date of the testing.}
\item{fips}{FIPS code for the locality (city or county).}
\item{locality}{Name of the locality.  We use the same abbreviation scheme
that VDH is using.  To join with other datasets, use the FIPS code.}
\item{district}{The health district that the locality belongs to.}
\item{population}{Total population of the locality.  Note that in Virginia
cities are not part of the counties that surround them.}
\item{ntest}{Estimated number of tests for that day in the locality.  Because
of proportional assignment, this will often not be an integer.}
}
}
\source{
Virginia Department of Health.  
\url{https://www.vdh.virginia.gov/coronavirus/}
}
\usage{
va_county_ntest
}
\description{
We estimate these values by scraping the number of tests by health district
from an infographic on the VDH website and assigning tests to the jurisdictions
within the district proportional to their share of the district's total population.
}
\details{
Unfortunately, each district's box in the infographic
is proportional to its share of the total, so the number is only readable for
the districts where the share of testing is large enough to make room to print
the district name and number of tests.  All of the other districts get lumped
together in a "Rest of Virginia" (ROVA) pseudo-district.  The districts for
which we have data include the most populous areas (NOVA, Richmond, and Virginia
Beach), as well as the Thomas Jefferson district, which includes the counties
surrounding Charlottesville.

An added complication is that we only have daily breakouts by district going back
to 2020-04-13.  Prior to that we have only the statewide total.  Therefore, for
that date and all prior dates we assign each district a share of the statewide
total proportional to its share of the cumulative totals on 2020-04-13, and then
we assign the shares within the district proportional to population share.

Because of the proportional assignment, the estimated number of tests will often
not be an integer.  In this dataset we just leave the fractional value in place.
For purposes of computing binomial or hypergeometric density functions, we round
such values \emph{up} to the next integer value.  Technically, this means that
our probability calculations are assuming a few more tests each day than were
actually performed; however, the error caused by this approximation is small
compared to the other approximations in use.
}
\keyword{datasets}
