% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{validate_params}
\alias{validate_params}
\title{Check that a parameter vector has no unknown parameters}
\usage{
validate_params(params)
}
\arguments{
\item{params}{Named list of parameter values}
}
\description{
Check against the names in the list of default parameters.
}
\keyword{internal}
