% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vdh_covid19}
\alias{vdh_covid19}
\title{Virginia Department of Health COVID-19 surveillance data}
\format{
Data frame with 5 columns:
\describe{
\item{Report.Date}{Date of the report}
\item{FIPS}{FIPS code for the locality}
\item{Locality}{Name of the locality}
\item{Health.District}{Health district the locality belongs to}
\item{Total.Cases}{Number of COVID-19 cases reported.  It's not documented in
the original source whether this is cumulative or for that day only, but the data
matches the figures given on the agency's website, which appears to be cumulative.}
}
}
\source{
Virginia Department of Health:\cr
\code{http://www.vdh.virginia.gov/content/uploads/sites/182/2020/03/VDH-COVID-19-PublicUseDataset-Cases.csv}
}
\usage{
vdh_covid19
}
\description{
Dataset with reported COVID-19 cases by date and locality.  It is not yet clear whether the
"Total Cases" reported are cumulative or new on that day.
}
\keyword{datasets}
