% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{wkagg}
\alias{wkagg}
\title{Aggregate model output to weekly resolution}
\usage{
wkagg(weekending, df, cols, aggfun = mean, nday = 7)
}
\arguments{
\item{weekending}{End dates for the weeks being calculated.}

\item{df}{Dataset to operate on}

\item{cols}{Columns to aggregate}

\item{aggfun}{Function to aggregate with (default is \code{mean})}

\item{nday}{Number of days in the "week".  Default is 7}
}
\description{
Compute the sum or average of the daily output from the model, for weeks
ending on dates specified in the output.  Aggregation will be performed only
for the requested columns; other columns will have their values at the
specified dates copied over unmodified.
}
