% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census-model.R
\name{census_model_output}
\alias{census_model_output}
\title{Compute input for the hospital census model from a filter model}
\usage{
census_model_output(fitdir, tfinal, mktmin = 0.1)
}
\arguments{
\item{fitdir}{Name of directory with saved filter fit objects}

\item{tfinal}{Max run time for the projections}

\item{mktmin}{Market share threshold.  Counties where we have a market share
below the threshold will not be run.}
}
\value{
Data frame suitable for input into census model. All case counts are
  market share weighted. Columns:
\describe{
\item{date}{Simulation date}
\item{time}{Simulation date expressed as days since Jan 01}
\item{scenario}{Ensemble member ID number}
\item{popmkt}{Market share weighted population}
\item{newCases}{Daily new cases}
\item{newSympto}{Daily new symptomatic cases}
\item{PopSympto}{Symptomatic population}
\item{PopInfection}{Infected population}
\item{fracInfection}{Infection prevalence.  Technically this is market share
  weighted, but you get the same answer whether you weight by market share or
  not.}
\item{PopCumulInfection}{Cumulative number of infections}
\item{fracCumulInfection}{Cumulative fraction of population infected.  This
  value is the same whether or not you weight by market share}
}
}
\description{
Run the ensemble members for all counties and aggregate their results,
weighted by market share.
}
\details{
If the model history is included, it will be prepended onto the projection.
If not, the output will start with the projection.
}
\section{TODO}{


* Refactor common code between this and \code{\link{project_filter_model}}.
* Add options for scenarios with transmissibility changes.
}

