% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagplots.R
\name{filter_model_diagnositc_plots}
\alias{filter_model_diagnositc_plots}
\title{Generate diagnostic plots for filter models}
\usage{
filter_model_diagnositc_plots(filterfit, simrun = NULL)
}
\arguments{
\item{filterfit}{A filter-fit structure returned by \code{\link{fit_filter}}}

\item{simrun}{Optional data frame or matrix of results from the run that generated
the input data.}
}
\value{
A list of ggplot objects
}
\description{
Generate diagnostic plots for filter models
}
