% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagplots.R
\name{filter_model_ensemble_plot}
\alias{filter_model_ensemble_plot}
\title{Plot traces for ensemble members in a filter model fit}
\usage{
filter_model_ensemble_plot(
  filterfit,
  what = "fi",
  Nplot = 50,
  highlight = NULL
)
}
\arguments{
\item{filterfit}{Filter-fit object}

\item{what}{Variable to plot.  Default is \code{fi}.}

\item{Nplot}{Number of ensemble members (randomly selected) to plot.  Default
is to plot 50.  If \code{NULL}, then plot all.}

\item{highlight}{List of ensemble member ids to overplot in a highlight
color.  Default is none.}
}
\description{
This plot requires that the history was kept in the filter fit.
}
