% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gen_likelihood}
\alias{gen_likelihood}
\title{Prepare a likelihood function for use in Bayesian calibration}
\usage{
gen_likelihood()
}
\value{
A function that takes a named list of model parameters and returns
a log-likelihood value.  See details for the parameters recognized.
}
\description{
Build a function that given a set of model parameters runs the model and 
computes the log-likelihood for the model.
}
\details{
The observational data is the daily county-by-county reports of new confirmed
COVID-19 cases.  We are using the New York Times dataset
(\code{https://github.com/nytimes/covid-19-data}), which is based on reports
from state departments of health, including the VDH.  For statistics on COVID-19
testing, we use the \code{\link{uvads_covid19}} dataset, which was collected
directly from VDH by the UVA data science team.

The model parameters currently recognized are:
\describe{
\item{T0}{(real > 0) The initial doubling time for the number of infected people}
\item{D0}{(real > 0) The initial average infection duration}
\item{A0}{(real > 0) The average incubation time}
\item{Ts}{(real > 0) The average time from infection to symptom onset.  Note that
this parameter is probably not identifiable with our current data.}
\item{day_zero}{(real > 0) The day on which the infection started in Fairfax County (not
necessarily when it was first observed), given in days since 1Jan2020.
This will be used to line up the observed dataset with model results.}
\item{b}{(real > 0) The testing bias factor.  That is, the positive test rate
divided by the true infection rate.  b can be different from 1 because of false positives
or because testing is targeted to people suspected of having the disease.  (Generally,
we expect b>1, but we don't require this.)}
}
For the time being, we only accept a single \code{day_zero} parameter, and 
all of the counties are delayed relative to Fairfax by a number of days equal
to the difference between their first observed case and the first case observed
in Fairfax.

To compute the likelihood, our assumption is that the output of the model 
represents an average infection rate.  The average observation rate is then
\deqn{\bar{N}_o = \bar{N}_I f_t b,} where \eqn{f_t} is the fraction of the total
population tested.  We then assume that the actual observations are distributed
as \deqn{N_o \sim Pois(\bar{N}_o).}
}
