% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{gen_post}
\alias{gen_post}
\title{Prepare a posterior log-pdf function for use in Bayesian calibratin}
\usage{
gen_post()
}
\description{
Prepare a posterior log-pdf function for use in Bayesian calibratin
}
