% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{get_obsdata}
\alias{get_obsdata}
\title{Prepare the observed data for use in the likelihood function}
\usage{
get_obsdata()
}
\value{
List of observations used in the likelihood function.  This includes
\describe{
  \item{obsdata}{Confirmed new cases by date and county.  Also includes 
  statewide fraction of people tested.}
  \item{fips_codes}{Lookup table of FIPS codes for the Locality names output by
  the model.}
  \item{default_parm_vals}{Default values for the parameters used to compare
  model output to observed data (the parameters used inside the model have their
  own defaults)}
}
}
\description{
Prepare the observed data for use in the likelihood function
}
\keyword{internal}
