% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagplots.R
\name{plt_modobs}
\alias{plt_modobs}
\title{Plot model predictions against data observations}
\usage{
plt_modobs(parms, scenarios = NULL, counties = NULL, default_parms = NULL)
}
\arguments{
\item{parms}{Parameters to run model for}

\item{scenarios}{Names of scenarios}

\item{counties}{Counties to include in the plot; if not specified, the top 4
by number of infections will be plotted.}

\item{default_parms}{Default values to use for parameters not specified in parms}

\item{obs}{Observed data as returned by \code{\link{get_obsdata}}}
}
\description{
Plot model outputs and observed data.
}
