% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagplots.R
\name{plt_projections}
\alias{plt_projections}
\title{Plot model projections for comparison}
\usage{
plt_projections(
  parms,
  scenarios,
  default_parms = c(day_zero = 30),
  tmax = 270,
  what = "newSympto",
  usedate = TRUE,
  counties = NULL,
  marketadjust = FALSE
)
}
\arguments{
\item{parms}{Matrix of combinations of parameters.}

\item{scenarios}{Vector of scenario names}

\item{default_parms}{Default values for parameters not mentioned in parms}

\item{tmax}{Maximum time to run to}

\item{what}{What output to plot.  Options are 
newCases, newSympto, PopSympto, PopInfection, PopCumulInfection, PopCumulFrac, popTotal}

\item{usedate}{If \code{TRUE}, display date on the x-axis; otherwise, display model
time.}

\item{counties}{If specified, filter to just the requested counties.  Otherwise,
include the whole state.}

\item{marketadjust}{If \code{TRUE}, adjust projections for UVAHS market share; 
otherwise, show raw totals.}
}
\description{
All of the results we plot here are adjusted for the UVA market fraction.
}
