% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagplots.R
\name{plt_projections}
\alias{plt_projections}
\title{Plot model projections for comparison}
\usage{
plt_projections(parms, scenarios, tmax = 270, what = "newSympto")
}
\arguments{
\item{parms}{Matrix of combinations of parameters.}

\item{scenarios}{Vector of scenario names}

\item{tmax}{Maximum time to run to}

\item{what}{What output to plot.  Options are 
newCases, newSympto, PopSympto, PopInfection, PopCumulInfection, PopCumulFrac, popTotal}
}
\description{
All of the results we plot here are adjusted for the UVA market fraction.
}
