% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior.R
\name{qprior}
\alias{qprior}
\title{Evaluate the quantile for the multidimensional df defined by the priors}
\usage{
qprior(p, hparms = list())
}
\arguments{
\item{p}{A vector of probabilities}

\item{hparms}{Named vector of hyperparameters, described in \code{\link{gen_post}}}
}
\description{
Given a vector of probabilities, find the corresponding quantile vector assuming
an independence copula and margins as described in \code{\link{gen_prior}}.
}
