% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{run_scenario}
\alias{run_scenario}
\title{Run a single hospital census scenario}
\usage{
run_scenario(tmax, params = list(), scenarioName = "HospCensus")
}
\arguments{
\item{tmax}{Maximum time (days) to run to}

\item{params}{List of parameter values, see details}

\item{scenarioName}{Name for the scenario; this will be copied into results}
}
\value{
Data frame with results (see details) over time
}
\description{
Run a hospital census scenario for a single set of model parameters.  Results for several
projected variables will be returned in a single data frame.
}
\details{
The results returned will be:
\itemize{
\item{New community infections in UVA referral base}
\item{New symptomatic community Infections in UVA referral base}
\item{New acute care admissions for UVA}
\item{New ICU admissions for UVA}
\item{New invasive mechanical ventilation admissions for UVA}
}

The parameters for the model are:
\describe{
\item{T0}{(scalar) Base doubling time.  Doubling time for number of cases when the number of infections
is small compared to the total population.}
\item{D0}{(scalar) Base recovery time.  Average base recovery time for an infected person.}
\item{(various time lags)}{Time after infection for various events (RTS for details)}
\item{(a whole slew of others)}{see definition of \code{param_defaults} for definitions}

}
}
