% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vaMyAgeBands}
\alias{vaMyAgeBands}
\title{Census data for Virginia localities}
\format{
Data frame with 8 columns
\describe{
\item{Locality}{Name of the locality}
\item{Total}{Total population}
\item{Age0to5}{Population ages 0--5}
\item{Age5to19}{Population ages 5--19}
\item{Age20to39}{Population ages 20--39}
\item{Age50to59}{Population ages 50--59}
\item{Age60to74}{Population ages 60--74}
\item{Age75plus}{Population over age 75}
}
}
\source{
Weldon Cooper Center (verify?)
}
\usage{
vaMyAgeBands
}
\description{
Dataset with total population and age breakdown for Virginia localities (counties,
county-equivalents, and the entire commonwealth).
}
\keyword{datasets}
